Shift tests for prepareModification to ModificationCommandTest from UpdateCommandTest
